+++ benchmarks/Main.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE CPP #-}
-
 module Main where
 
 import Bench.AddMod qualified
@@ -11,12 +9,14 @@ import Bench.PowMod qualified
 import Bench.RepeatWithIndex qualified
 import Bench.RepeatWithoutIndex qualified
 import Bench.SwapDupe qualified
+import Bench.Vector.ConcatMapM qualified
+import Bench.Vector.IConcatMapM qualified
 import Criterion.Main
 
 main :: IO ()
 main =
   defaultMain
-    -- TODO: generate criterion graph by benchmark group?
+    -- TODO: Filter benchmark cases with command line arguments. Currently I'm filtering benchmarks by hand.
     [ Bench.MulMod.benches,
       Bench.ModInt.benches,
       Bench.AddMod.benches,
@@ -25,5 +25,7 @@ main =
       Bench.Montgomery64.benches,
       Bench.RepeatWithIndex.benches,
       Bench.RepeatWithoutIndex.benches,
-      Bench.SwapDupe.benches
+      Bench.SwapDupe.benches,
+      Bench.Vector.ConcatMapM.benches,
+      Bench.Vector.IConcatMapM.benches
     ]